Cache inode open/close/rdwr/commit
This commit includes the changes to cache_inode_open,
cache_inode_close, cache_inode_rdwr, and cache_inode_commit as well as
their support and callers.  Primarily, this change involves the use of
entry->content_lock to protect the write buffer and file descriptor
and the removing max FD management from cache_indoe_close.  Also in
this patch are the removal of fsal_seek_t from cache_inode interfaces
(it is still used to call FSAL_read and FSAL_write, for now),
normalizing of the stability parameters to cache_inode_rdwr and
cache_inode_commit (It was originally a uint64_t with three #define'd
values, except some functions treated it as a boolean and set it to or
compared it against TRUE and FALSE.  It is now an enumeration with the
defined values used consistently), the removal of
cache_content_io_direction_t (it had no reason to be disjoint from
cache_inode_io_direction_t), and movement of Max_Fd and fd_cache from
thne client structure to the gc_policy structure.

This patch represents a snapshot of work in progress and is not
suitable for use.  It has not been tested, and forthcoming patches
will be required to add a state lock (and appropriate management) and
the example FD reaper in cache_inode_lru.