Add state_lock to cache_entry_t
In this patch, create a separate state_lock in cache_entry_t.  My
original idea was to have state protected by the content_lock (which,
in regular files, protects a cached file descriptor.)  However, it is
much better if we can release the lock on the file descriptor while
still maintaining the lock on NFS state for the duration of an
operation.

Additionally, I moved the state_list from inside the file fork of the
object union to the main object, since state can exist on non-regular
file objects (NFSv4.1 directory delegations, specifically.)