Preliminary LRU L2 example
This commit implements a preliminary example of the use of L2.  When
the LRU thread wakes, if the number of FDs is greater than the high
water mark, it picks LRU_WORK_PER_WAKE elements from each queue,
checks them for file descriptors and frees them if appropriate, and
moves each checked entry to L2.

This is an unstable point in development.  Further coding, testing,
and fixing are required.  There are known bugs.