Deprecate USE_SHARED_FSAL configuration option
Remove USE_SHARED configuration option from code in preparation
for the new fsal api and infrastructure for multiple loadable
fsal modules.

Padding to fsal specific structures in */fsal_types.h are
unconditionally part of the union.  This makes all instances of
these structs the same size but there is no checking for oversize.

Signed-off-by: Jim Lieb <jlieb@panasas.com>