Remove nbwanted parameter
The cache_inode_readdir function will provide results until the
callback signals that the result structure is full and an entry has
not been added and return eod only if the last entry has been
signalled as in the result.  This simplifies both returning EOF to
clients and the handling of . and .. entries for NFSv3 and NFSv2.