Persistent offsets in dirent AVL
Make offsets (ie, cookies) persistent.  Any cookie allocated since the
full population of a dirent cache is preserved, the associated dirent is
marked deleted.  If a deleted offset is ever hashed later, the slot is reused.

The strategy is to move deleted entries to an alternate AVL.  If a READDIR
request is made for an offset which is no longer active, entries are returned
starting from the least upper bound of offset in the -active- AVL.  Unit
tests for the new avltree_sup and former avltree_inf methods are added.

A hard upper bound of 64K deleted entries per cache entry is enforced.  Further,
we expect that deleted entries may be purged after periods of inactivity.
(Persistent cookies are a convience for broken clients.)

Conflicts:

	src/Cache_inode/cache_inode_avl.c
	src/Cache_inode/cache_inode_misc.c
	src/Cache_inode/cache_inode_readdir.c
	src/Cache_inode/cache_inode_renew_entry.c
	src/include/cache_inode.h