Do not gratuitously destroy memory in readdir
The readdir callback functions were tracking maxcount (the total
number of bytes the client is willing to accept in a response) and
allocating space for filenames/NFSv4 handles/Attributes dynamically.
However it was not checking the size of the allocated entry2, entry3,
entryplus3, or entry4 array.  This bug was not apparently until
recently, since the nbwanted parameter kept cache_inode from ever
calling the callback with more than the requested number of entries.