NFSv3 - don't trash memory in readdir
Just like it was done in readdirplus case if the number of
entries allocated can be limited artificially
to 50 or less then size of buffer alone cannot be used to
decide when to stop - you need the "artificial" value too.

Also count the space used by the buffer very carefully.

And make sure buffers which are meant to be zero-filled,
such as those used for holding array or entries, are filled
with zeros.

Signed-off-by: Max Matveev <makc@gmx.co.uk>