NFS4.0 share reservation: Add the proper handling of open, open upgrade, open downgrade, close, and expiry.
No code changes in NFS4.1 share reservation.

Add the ref counted share state to cache_entry_t. This is for quickly
checking share conflict and easily updating share state.

Add the bitmap of previously seen share access and deny to share
state. This is for checking valid open downgrade.

Add the capability flags to FSAL to indicate the share reservation
support.

The typical work flow of share state update is (a) updating ref counted
share state, (b) calculating the union of file share states based on the
change, (c) push it down to FSAL if it is different than current, (d)
update share bits in state.

Tests done with GPFS and VFS FSAL
  - build
  - cthon v3 and v4
  - pynfs 4.0 open, openconfirm, opendowngrade
  - NFSv4.1: basic file access