Introduction of nfs4_sanity_check_FH and some very basic code refactoring for NFSv4.
Same as the previous patch, but now for nfsv4 files.

Introducing nfs4_sanity_check_FH to the parts of the code where it was applicable.

Also some code refactoring. Removing some intermediate variables, used to store
the return value of some functions. In those cases, removing the variable
assignment from the 'if' statement. Some symmetry reorg for a better look :-)
And some typos fix and french translations.

Signed-off-by: Leonardo Tonetto <tonetto@linux.vnet.ibm.com>