Make cache_inode_new_entry behave properly on finding a dead entry
In this changeset, we modify cache_inode_new_entry to replace any dead
entry it finds in the hash table.  We add a new function,
HashTable_DelSafe that searches for a key and checks that values match
before deletion, and use it in cache_inode_kill_entry and
cache_inode_clean_internal.  The effect is such that these functions
will only delete a hash table entry if it has not been replaced by
cache_inode_new_entry.  (The original cache entry that was replaced
will still be deconstructed, however.)