Added a new log function to print debug info during specific log levels.
The new function will trigger debug info like backtraces, global
variables, etc. to be printed after any log message that is less then or
equal to the log level of LOG_MESSAGE_DEBUGINFO, a new component.
LOG_MESSAGE_DEBUGINFO is set to NIV_NULL by default, so it has to be
explicitly enabled.