Fix config parsing so long var_name and var_value strings are reported
Ganesha has a limit of 1023 characters for config option names and
value strings, but that limit was not actually reported or enforced.
The parser just captures 1024 bytes of the string.

This patch makes Ganesha check if strings are terminated within
1024 characters, and if not returns an error. config_GetKeyValue
will now return -2 if the var_name string is more than 1023
characters, and it will terminate the string so callers can
report the error. It will return -3 if the var_value string
is too long, and again terminate the string.

Signed-off-by: Frank Filz <ffilz@us.ibm.com>