Remove redundant state/share conflict tests
The old code did THREE share conflict checks on every NFS v4 open...

The first check (which is the one retained) is a call to
state_share_check_conflict after doing a permission check for open (to
dispose of share conflicts quickly).

The second check was from state_add_impl. This check is actually insidious
because EVERY new state, whether it is a share reservation or not is checked
for conflict with every other stateid on the file. Of course the only ACTUAL
conflicts checked are between share reservations...

The final check was from either state_share_add or state_share_upgrade.

Note that we do NOT need to check differently on upgrade due to this text in RFC
3530bis in Section 9.9. Share Reservations:

This checking of share reservations on OPEN is done with no exception
for an existing OPEN for the same open-owner.

I also added a todo comment indicating that the anonymous IO temporary share
implemented with state_share_anonymous_io_start and
state_share_anonymous_io_done needs to distinguish v3 and v4 anonymouse I/O.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>